requirements.txt
torch==1.8.0
torchvision==0.9.0
opencv-python
numpy==1.16.0
tqdm
scipy
pathlib
pandas
Pillow
argparse
json
pickle5
open3d
pytorch-lightning
seaborn
matplotlib
torchmetrics
scikit-learn